+++ output/colors_test.go
@@ -1,9 +1,9 @@
 package output
 
 import (
-	"testing"
 	"github.com/jedib0t/go-pretty/text"
 	log "github.com/sirupsen/logrus"
+	"testing"
 )
 
 func TestColors(t *testing.T) {
@@ -19,8 +19,12 @@ func TestColors(t *testing.T) {
 	YellowBold("Yellow Bold message")
 	Cyan("Cyan message")
 	CyanBold("Cyan Bold message")
-	White("White message")
-	WhiteBold("White Bold message")
+	HiWhite("HiWhite message")
+	HiWhiteBold("HiWhiteBold Bold message")
+	HiYellow("HiYellow message")
+	HiYellowBold("HiYellowBold Bold message")
+	HiMagenta("HiMagenta message")
+	HiMagentaBold("HiMagentaBold Bold message")
 	Message("Generic message", text.Colors{text.Bold, text.BgGreen})
 	//ErrorExit("TestColors", "Error message bye bye...")
 }